Adding csproj and solution file for microbuild signing
The csproj and sln files will not be connected to anything else in the repo, and will be effectively inert and ignorable to all other users.
But we are creating a signing microbuild definition, which when run will msbuild the csproj file, and that build create a zip file containing signed versions of the JS files in the extension.  This signed zip will be used to produce the nuget file that will eventually be automatically inserted into the VS willow tree and setup.